+++ src/trait_/debug.rs
@@ -1,21 +1,28 @@
 //! [`Debug`](trait@std::fmt::Debug) implementation.
 
+use std::ops::Deref;
+
 use proc_macro2::TokenStream;
 use quote::quote;
 
-use crate::{Data, DeriveTrait, DeriveWhere, Item, SimpleType, SplitGenerics, TraitImpl};
+use crate::{
+	util, Data, DeriveTrait, DeriveWhere, Item, SimpleType, SplitGenerics, Trait, TraitImpl,
+};
 
-/// Dummy-struct implement [`Trait`](crate::Trait) for
-/// [`Debug`](trait@std::fmt::Debug).
+/// [`TraitImpl`] for [`Debug`](trait@std::fmt::Debug).
 pub struct Debug;
 
 impl TraitImpl for Debug {
-	fn as_str(&self) -> &'static str {
+	fn as_str() -> &'static str {
 		"Debug"
 	}
 
-	fn default_derive_trait(&self) -> DeriveTrait {
+	fn default_derive_trait() -> DeriveTrait {
 		DeriveTrait::Debug
+	}
+
+	fn path(&self) -> syn::Path {
+		util::path_from_strs(&["core", "fmt", "Debug"])
 	}
 
 	fn build_signature(
@@ -23,7 +30,6 @@ impl TraitImpl for Debug {
 		_derive_where: &DeriveWhere,
 		_item: &Item,
 		_generics: &SplitGenerics<'_>,
-		_trait_: &DeriveTrait,
 		body: &TokenStream,
 	) -> TokenStream {
 		quote! {
@@ -35,23 +41,16 @@ impl TraitImpl for Debug {
 		}
 	}
 
-	fn build_body(
-		&self,
-		_derive_where: &DeriveWhere,
-		trait_: &DeriveTrait,
-		data: &Data,
-	) -> TokenStream {
+	fn build_body(&self, _derive_where: &DeriveWhere, data: &Data) -> TokenStream {
 		let self_pattern = &data.self_pattern();
 		let debug_name = data.ident.to_string();
 
 		match data.simple_type() {
 			SimpleType::Struct(_) => {
-				let self_ident = data.iter_self_ident(**trait_);
-				let debug_fields = data
-					.iter_field_ident(**trait_)
-					.map(|field| field.to_string());
+				let self_ident = data.iter_self_ident(**self);
+				let debug_fields = data.iter_field_ident(**self).map(|field| field.to_string());
 
-				let finish = if data.any_skip_trait(**trait_) {
+				let finish = if data.any_skip_trait(**self) {
 					quote! { finish_non_exhaustive }
 				} else {
 					quote! { finish }
@@ -66,7 +65,7 @@ impl TraitImpl for Debug {
 				}
 			}
 			SimpleType::Tuple(_) => {
-				let self_ident = data.iter_self_ident(**trait_);
+				let self_ident = data.iter_self_ident(**self);
 
 				quote! {
 					#self_pattern => {
@@ -81,5 +80,13 @@ impl TraitImpl for Debug {
 			}
 			SimpleType::Union => unreachable!("unexpected trait for union"),
 		}
+	}
+}
+
+impl Deref for Debug {
+	type Target = Trait;
+
+	fn deref(&self) -> &Self::Target {
+		&Trait::Debug
 	}
 }