+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- `no_drop` item-level option to `ZeroizeOnDrop` which does not implement
+  `Drop` but instead only asserts that every field implements `ZeroizeOnDrop`.
+
 ## [1.4.0] - 2025-05-01
 
 ### Added